#878: 파일이동
- 가이드씬 모듈에서 CommonViews 모듈로 이동합니다.
